# fileservice: optimize fifocache.DataCache.EnsureNBytes