--- a/pkg/fileservice/local_fs.go
+++ b/pkg/fileservice/local_fs.go
@@ -113,7 +113,11 @@ func NewLocalFS(
 
 func (l *LocalFS) AllocateCacheData(size int) fscache.Data {
 	if l.memCache != nil {
-		l.memCache.cache.EnsureNBytes(size)
+		l.memCache.cache.EnsureNBytes(
+			size,
+			// evict at least 1/100 capacity to reduce number of evictions
+			int(l.memCache.cache.Capacity()/100),
+		)
 	}
 	return DefaultCacheDataAllocator().AllocateCacheData(size)
 }