## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue #19227 

## What this PR does / why we need it:
new:
BenchmarkEnsureNBytesAndSet-24    	35649835	        34.06 ns/op
BenchmarkEnsureNBytesAndSet-24    	35766162	        33.51 ns/op
BenchmarkEnsureNBytesAndSet-24    	37361575	        33.85 ns/op
BenchmarkEnsureNBytesAndSet-24    	36150702	        34.08 ns/op
BenchmarkEnsureNBytesAndSet-24    	36303979	        33.88 ns/op

old:
BenchmarkEnsureNBytesAndSet-24    	28314853	        40.74 ns/op
BenchmarkEnsureNBytesAndSet-24    	31413367	        42.83 ns/op
BenchmarkEnsureNBytesAndSet-24    	28103589	        42.30 ns/op
BenchmarkEnsureNBytesAndSet-24    	33316714	        39.27 ns/op
BenchmarkEnsureNBytesAndSet-24    	30268882	        39.39 ns/op